Use `JsonDocument` instead of `Dictionary<string, object>`

+++ src/Isbm2Client.Test/TestObject.cs
@@ -8,7 +8,8 @@ namespace Isbm2Client.Test
 {
     public class TestObject
     {
-        public int[] Numbers { get; set; } = null!;
+        // TODO: Needs to be set back to int[] issue for server has been resolved
+        public double[] Numbers { get; set; } = null!;
         public string Text { get; set; } = "";
         public Dictionary<string, double> Weather { get; set; } = null!;
     }
